refactor: remove SyncRepositoriesController and simplify HTTP server setup
