+++ src/http-server/controllers/GetRoomListController.ts
@@ -1,12 +1,12 @@
 import { Request, Response } from "express";
 
 import RoomList from "../../edopro/room/infrastructure/RoomList";
-import MercuryRoomList from "../../mercury/room/infrastructure/MercuryRoomList";
+import YGOProRoomList from "../../mercury/room/infrastructure/YGOProRoomList";
 
 export class GetRoomListController {
 	run(_req: Request, response: Response): void {
 		const rooms = RoomList.getRooms().map((room) => room.toPresentation());
-		const mercuryRooms = MercuryRoomList.getRooms().map((room) => room.toPresentation());
-		response.status(200).json({ rooms: [...mercuryRooms, ...rooms] });
+		const ygoproRooms = YGOProRoomList.getRooms().map((room) => room.toPresentation());
+		response.status(200).json({ rooms: [...ygoproRooms, ...rooms] });
 	}
 }
