+++ src/shared/room/application/RoomFinder.ts
@@ -1,5 +1,5 @@
 import RoomList from "../../../edopro/room/infrastructure/RoomList";
-import MercuryRoomList from "../../../mercury/room/infrastructure/MercuryRoomList";
+import MercuryRoomList from "../../../mercury/room/infrastructure/YGOProRoomList";
 import { YgoRoom } from "../domain/YgoRoom";
 
 export class RoomFinder {
